+++ src/isal/__init__.py
@@ -17,19 +17,14 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-from typing import Optional
 
-try:
-    from . import _isal
-    ISAL_MAJOR_VERSION: Optional[int] = _isal.ISAL_MAJOR_VERSION
-    ISAL_MINOR_VERSION: Optional[int] = _isal.ISAL_MINOR_VERSION
-    ISAL_PATCH_VERSION: Optional[int] = _isal.ISAL_PATCH_VERSION
-    ISAL_VERSION: Optional[str] = _isal.ISAL_VERSION
-except ImportError:  # isa-l.h not available on windows
-    ISAL_MAJOR_VERSION = None
-    ISAL_MINOR_VERSION = None
-    ISAL_PATCH_VERSION = None
-    ISAL_VERSION = None
+from . import _isal
+
+ISAL_MAJOR_VERSION: int = _isal.ISAL_MAJOR_VERSION
+ISAL_MINOR_VERSION: int = _isal.ISAL_MINOR_VERSION
+ISAL_PATCH_VERSION: int = _isal.ISAL_PATCH_VERSION
+ISAL_VERSION: str = _isal.ISAL_VERSION
+
 
 __all__ = [
     "ISAL_MAJOR_VERSION",